+++ src/app/types/authentication.ts
@@ -1,0 +1,14 @@
+export interface Login {
+  login: string;
+  password: string;
+  returnToken?: boolean
+}
+
+export interface Token {
+  email: string;
+  id: string;
+  login: string;
+  role: Array<string>;
+  token: string;
+  type: string;
+}
